+++ ory-kratos-sdk/mutations/update-login-flow.ts
@@ -8,7 +8,7 @@ import type {
 } from "@ory/client";
 import { TanstackError } from "@saflib/sdk";
 import { getKratosFrontendApi } from "../kratos-client.ts";
-import { invalidateKratosSessionQueries } from "../kratos-session.ts";
+import { invalidateKratosSessionQueries } from "../queries/kratos-session.ts";
 import { BrowserRedirectRequired } from "../flow-results.ts";
 import {
   LoginFlowFetched,
