+++ ory-kratos-sdk/mutations/update-recovery-flow.ts
@@ -7,7 +7,7 @@ import type {
 } from "@ory/client";
 import { TanstackError } from "@saflib/sdk";
 import { getKratosFrontendApi } from "../kratos-client.ts";
-import { invalidateKratosSessionQueries } from "../kratos-session.ts";
+import { invalidateKratosSessionQueries } from "../queries/kratos-session.ts";
 import { BrowserRedirectRequired } from "../flow-results.ts";
 import {
   RecoveryFlowFetched,
